requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
ddt==1.7.2
et_xmlfile==2.0.0
filelock==3.16.1
fsspec==2024.12.0
HessianPy==1.0.4
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
Logbook==1.8.0
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pluggy==1.5.0
protocol==0.37
pydicom==3.0.1
pyrequest==0.5
pytest==8.3.4
python-hessian==1.2.0
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sympy==1.13.3
torch==2.2.2
typing_extensions==4.12.2
urllib3==2.3.0
xlrd==1.2.0
xlwt==1.3.0
